requirements.txt

python-telegram-bot==22.3
gspread==6.2.1
oauth2client==4.1.3
nest_asyncio==1.6.0
Flask==2.3.3
